+++ backend/controllers/products.js
@@ -1,10 +1,15 @@
 const db = require('../models')
 const Product = db.product
 const { paginationFromTo } = require('../helper/pagination')
+const { category } = require('../models')
 
 exports.AddProduct = async (req, res, next) => {
     const product = await Product.create(req.body)
-    res.send(product)
+    if (product) {
+        res.status(200).json(`Product added successfully`)
+    } else {
+        res.status(400).json(product)
+    }
 }
 
 exports.GetAllProduct = async (req, res, next) => {
@@ -25,29 +30,57 @@ exports.GetAllProduct = async (req, res, next) => {
     const { from, to } = req.query
     const params = (!from && !to) ? {} : paginationFromTo({ from, to })
     console.log({ params });
-    const products = await Product.findAndCountAll(params)
-    res.send(products)
+    const products = await Product.findAndCountAll({
+        ...params, 
+        order: [
+        ['productID', 'ASC']]
+    })
+    // res.send(products)
+    if (products) {
+        res.status(200).send(products)
+    } else {
+        res.status(400).json(category)
+    }
 }
 
 exports.GetProductByID = async (req, res, next) => {
     const { id } = req.params
     const product = await Product.findByPk(id)
-    res.send(product)
+    if (product) {
+        res.status(200).send(product)
+    } else {
+        res.status(400).json(product)
+    }
 }
 
 exports.UpdateProduct = async (req, res, next) => {
     const { id } = req.params
-    const product = await Product.update(
-        { ...req.body },
-        { where: { productID: id } }
-    )
-    res.send(product)
+    const { categoryID } = req.body
+    const categoryExists = await Category.findByPk(categoryID)
+
+    let product = null
+    if (categoryExists) {
+        product = await Product.update(
+            { ...req.body },
+            { where: { productID: id } }
+        )
+    }
+    if (product) {
+        res.status(200).json(`Product updated successfully`)
+    } else {
+        res.status(400).json(product)
+    }
 }
 
 exports.DeleteProduct = async (req, res, next) => {
     const { id } = req.params
-    await Product.destroy({
-        where: { productID: id }
-    })
-    res.sendStatus(200)
+    const product = await Product.findByPk(id)
+    if (product) {
+        await Product.destroy({
+            where: { productID: id }
+        })
+        res.status(200).json(`Product deleted successfully`)
+    } else {
+        res.status(400).json(`Product does not exist`)
+    }
 }
